replace onAttachmentContextClicked Experiment
This also removes the setOptions Experiment, because it was only used to make contextInfo available to the Experiment, which is no longer needed.

It was not used to set logenabled. which is hardcoded to true. If you want to make it the follow the user defined option, we will find a different way.